add important desktop file back

+++ debian-install.sh
@@ -10,7 +10,7 @@ check_exit() {
 }
 
 # Update system
-sudo apt update && sudo apt full-upgrade
+sudo apt update && sudo apt upgrade
 check_exit "System update and upgrade"
 
 # Install nala
@@ -30,7 +30,7 @@ sleep 1
 sudo nala install git network-manager nm-connection-editor ufw bluez \
   tlp powertop upower \
   pipewire pipewire-pulse wireplumber \
-  build-essetial man-db brightnessctl curl wget gawk aria2  btop \
+  build-essential man-db brightnessctl curl wget gawk aria2  btop \
   efibootmgr timeshift
 check_exit "Installing system packages"
 
@@ -39,14 +39,14 @@ sudo nala install sway swaybg sway-notification-center swayidle swaylock \
   i3status autotiling foot foot-themes foot-terminfo wmenu wl-clipboard \
   xdg-user-dirs xdg-desktop-portal xdg-desktop-portal-wlr xwayland \
   nwg-look nwg-displays wev wlr-randr wl-mirror \
-  fonts-noto fonts-noto-cjk ttf-ms-fonts
+  fonts-noto fonts-noto-cjk ttf-mscorefonts-installer
 check_exit "Installing sway packagages"
 
 # Install more tools
 sudo nala install zsh stow starship zoxide eza bat fd-find ncdu fzf ripgrep tmux pipx rustc cargo npm \
   firefox qutebrowser flatpak zip unzip 7zip neovim tree-sitter-cli \
   gammastep zoxide slurp grim imv zathura \
-  fastfetch tree lazygit rocks git-lfs
+  fastfetch tree lazygit luarocks git-lfs
 check_exit "Installing additional tools"
 
 xdg-user-dirs-update
@@ -62,9 +62,15 @@ pipx install tldr
 check_exit "Installing tldr"
 
 # Backup existing configs
-mv "$HOME/.bashrc" "$HOME/.bashrc.bak"
-mv "$HOME/.zshrc" "$HOME/.zshrc.bak"
-mv "$HOME/.oh-my-zsh" "$HOME/.oh-my-zsh.bak"
+if [ -e "$HOME/.bashrc" ]; then
+ mv "$HOME/.bashrc" "$HOME/.bashrc.bak"
+fi
+if [ -e "$HOME/.zshrc" ]; then
+ mv "$HOME/.zshrc" "$HOME/.zshrc.bak"
+fi
+if [ -e "$HOME/.oh-my-zsh" ]; then
+ mv "$HOME/.oh-my-zsh" "$HOME/.oh-my-zsh.bak"
+fi
 check_exit "Backing up existing configs"
 
 # Install zsh and plugins
@@ -82,7 +88,7 @@ check_exit "Removing old zshrc"
 # Download and install JetBrains Mono Nerd Fonts
 mkdir -p "$HOME/.fonts"
 cd "$HOME"
-aria2c --out "Downloads/" https://release-assets.githubusercontent.com/github-production-release-asset/27574418/618d36f5-7bcc-4e03-a153-30d18952aa14?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-03-18T18%3A31%3A30Z&rscd=attachment%3B+filename%3DJetBrainsMono.zip&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-03-18T17%3A30%3A39Z&ske=2026-03-18T18%3A31%3A30Z&sks=b&skv=2018-11-09&sig=G
+#aria2c --out "Downloads/JetBrainsMono.zip" https://release-assets.githubusercontent.com/github-production-release-asset/27574418/618d36f5-7bcc-4e03-a153-30d18952aa14?sp=r&sv=2018-11-09&sr=b&spr=https&se=2026-03-18T18%3A31%3A30Z&rscd=attachment%3B+filename%3DJetBrainsMono.zip&rsct=application%2Foctet-stream&skoid=96c2d410-5711-43a1-aedd-ab1947aa7ab0&sktid=398a6654-997b-47e9-b12b-9515b896b4de&skt=2026-03-18T17%3A30%3A39Z&ske=2026-03-18T18%3A31%3A30Z&sks=b&skv=2018-11-09&sig=G
 unzip "$HOME/Downloads/JetBrainsMono.zip" -d "$HOME/.fonts/"
 check_exit "Downloading and installing JetBrainsMono fonts"
